Added Support for HTTPS
The server now supports HTTPS running as a second server and on a different port. It requires the use of a JKS java key store file and includes instructions for converting more popular PEM, PKCS12 and PKCS7 certificates. This allows for a key store connected to a valid certificate authority and completing a valid certificate chain to be provided for encoding by the origin server. It should be noted that, except for special cases, for the purposes of testing an address will not match any of those for which the certificate provides encoding through a registered certificate authority or through the prior validating point in the certificate chain and will therefore not be excepted by most popular and secure browsers as a valid certificate. This will not, however, disable the encryption so proceeding to the webpage will allow you to test on a secure connection.