Standardized transmission character sets
Due to the fact that operating systems disagree on default encoding some HTML symbol codes would result in improper or multiple characters when the page was served from a UNIX based system such as OSX, Linux and Ubuntu. The system now explicitly forces the JVM to use UTF-8 instead of varied default encodings.