Made improvements to encoding
Added support for Deflate encoding instead of just returning empty pages when it was requested because I forgot to add that before. Organized compression options with an enumerator for increased ease of implementation in the future for adding Brotli or other encoding methods. Resolved an error that prevented the server from returning any content or throwing a visible error when the client did not send an Accept-Encoding header.